requirements.txt
diffusers==0.32.1
pillow==11.0.0
torch==2.5.1
transformers==4.47.1